Fix clipboard handling in safari (#6917)
* fix clipboard handling in safari

closes #6911, #6909


* [autofix.ci] apply automated fixes

* update dependencies

---------
