+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   ([#6876](https://github.com/mitmproxy/mitmproxy/pull/6876), @errorxyz)
 * Add an API to parse HTTPS records from DNS RDATA.
   ([#6884](https://github.com/mitmproxy/mitmproxy/pull/6884), @errorxyz)
+* Fix flow export in mitmweb for Safari
+  ([#6917](https://github.com/mitmproxy/mitmproxy/pull/6917), @mhils, @canyesilyurt)
 * Releases now come with a Sigstore attestations file to demonstrate build provenance.
   ([f05c050](https://github.com/mitmproxy/mitmproxy/commit/f05c050f615b9ab9963707944c893bc94e738525), @mhils)
