+++ web/package.json
@@ -8,7 +8,7 @@
         "prettier": "prettier --write ."
     },
     "dependencies": {
-        "@popperjs/core": "^2.9.3",
+        "@popperjs/core": "^2.11.8",
         "bootstrap": "^3.4.1",
         "classnames": "^2.3.1",
         "codemirror": "^5.62.3",
@@ -29,7 +29,7 @@
         "@testing-library/jest-dom": "^5.14.1",
         "@testing-library/react": "^12.0.0",
         "@testing-library/user-event": "^13.2.1",
-        "@types/jest": "^27.0.1",
+        "@types/jest": "^29.5.12",
         "@types/redux-mock-store": "^1.0.3",
         "esbuild": "^0.12.21",
         "esbuild-jest": "^0.5.0",
@@ -43,7 +43,8 @@
         "gulp-plumber": "^1.2.1",
         "gulp-replace": "^1.1.3",
         "gulp-sourcemaps": "^3.0.0",
-        "jest": "^27.0.6",
+        "jest": "^29.7.0",
+        "jest-environment-jsdom": "^29.7.0",
         "jest-fetch-mock": "^3.0.3",
         "prettier": "2.8.4",
         "react-test-renderer": "^17.0.2",
