+++ web/src/js/flow/export.ts
@@ -1,14 +1,22 @@
-import { runCommand } from "../utils";
+import { copyToClipboard, runCommand } from "../utils";
 import { Flow } from "../flow";
-import { copyToClipboard } from "../contrib/clipboard";
 
 export const copy = async (flow: Flow, format: string): Promise<void> => {
-    let ret = await runCommand("export", format, `@${flow.id}`);
-    if (ret.value) {
-        await copyToClipboard(ret.value);
-    } else if (ret.error) {
-        alert(ret.error);
-    } else {
-        console.error(ret);
+    // Safari: We need to call copyToClipboard _right away_ with a promise,
+    // otherwise we're loosing user intent and can't copy anymore.
+    let formatted = (async () => {
+        let ret = await runCommand("export", format, `@${flow.id}`);
+        if (ret.value) {
+            return ret.value;
+        } else if (ret.error) {
+            throw ret.error;
+        } else {
+            throw ret;
+        }
+    })();
+    try {
+        await copyToClipboard(formatted);
+    } catch (err) {
+        alert(err);
     }
 };
